requirement.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
nose==1.3.7
python-dateutil==2.7.2
PyYAML==3.12
six==1.11.0
svn==0.3.45
Werkzeug==0.14.1
